requirements.txt
fastapi
uvicorn[standard]
openai-whisper
anthropic
python-multipart
jinja2
python-dotenv
